refactor(layout): 抽象出代码编辑器，布局设置，状态监听器的相关逻辑
顺便修改一下CustomList新增按钮的样式
